+++ packages/@leight/query/src/$export/IUseMutationResult.ts
@@ -1,1 +1,0 @@
-export {type IUseMutationResult} from "../api/IUseMutationResult";