+++ packages/@leight/query/src/$export/IUseQueryResult.ts
@@ -1,1 +1,0 @@
-export {type IUseQueryResult} from "../api/IUseQueryResult";