+++ packages/@leight/query/src/api/IUseMutationResult.ts
@@ -1,4 +1,0 @@
-import {type UseMutationResult} from "@tanstack/react-query";
-import {type IErrorResponse}    from "../schema/ErrorResponseSchema";
-
-export type IUseMutationResult<TRequest, TResponse> = UseMutationResult<TResponse, IErrorResponse, TRequest>;