+++ packages/@leight/query/src/api/IUseQueryResult.ts
@@ -1,4 +1,0 @@
-import {type UseQueryResult} from "@tanstack/react-query";
-import {type IErrorResponse} from "../schema/ErrorResponseSchema";
-
-export type IUseQueryResult<TResponse> = UseQueryResult<TResponse, IErrorResponse>;